// src/app/shared/model/uc-address-obj.model.ts
export class UcAddressObj {
        Addr: string;
        AreaCode4: string;
        AreaCode3: string;
        AreaCode2: string;
        AreaCode1: string;
        City: string;
        PhnArea1: string;
        Phn1: string;
        PhnExt1: string;
        PhnArea2: string;
        Phn2: string;
        PhnExt2: string;
        PhnArea3: string;
        Phn3: string;
        PhnExt3: string;
        Notes: string;
        FaxArea: string;
        Fax: string;
        MrHouseOwnershipCode: string;
        SubZipcode: string;
        StaySince: string;
        StayLength: number;
        RowVersion: string;

        constructor() {
                this.Addr = "";
                this.AreaCode4 = "";
                this.AreaCode3 = "";
                this.AreaCode2 = "";
                this.AreaCode1 = "";
                this.City = "";
                this.PhnArea1 = "";
                this.Phn1 = "";
                this.PhnExt1 = "";
                this.PhnArea2 = "";
                this.Phn2 = "";
                this.PhnExt2 = "";
                this.PhnArea3 = "";
                this.Phn3 = "";
                this.PhnExt3 = "";
                this.FaxArea = "";
                this.Fax = "";
                this.MrHouseOwnershipCode = "";
                this.SubZipcode = "";
                this.StaySince = "";
                this.StayLength = 0;
                this.RowVersion = "";
        }
}